Request Ride Finished
I got most of the Request Ride done, to include adding the button for
GPS.  The yes and no questions are not working quite right but I will
get to them later.  I also have it locked on portrait mode.  I also
haven't combined the RequestRideSubmit and the Form_Fill.class.  If you
want to keep trying out the connection use the "test connection to
server" button on the homescreen.